Split Board abstraction
