+++ pkg/abstraction/board.go
@@ -1,0 +1,14 @@
+package abstraction
+
+type BoardEvent string
+
+type BoardNotification interface {
+	Event() BoardEvent
+}
+
+type BoardId uint16
+
+type Board interface {
+	Id() BoardId
+	Notify(BoardNotification)
+}
