+++ pkg/vehicle/constructor.go
@@ -4,7 +4,7 @@ import "github.com/HyperloopUPV-H8/h9-backend/pkg/abstraction"
 
 func New() Vehicle {
 	return Vehicle{
-		boards: make(map[BoardId]Board),
+		boards: make(map[abstraction.BoardId]abstraction.Board),
 	}
 }
 
@@ -12,15 +12,15 @@ func (vehicle *Vehicle) SetBroker(broker abstraction.Broker) {
 	vehicle.broker = broker
 }
 
-func (vehicle *Vehicle) AddBoard(board Board) {
+func (vehicle *Vehicle) AddBoard(board abstraction.Board) {
 	vehicle.boards[board.Id()] = board
 }
 
-func (vehicle *Vehicle) RemoveBoard(board Board) {
+func (vehicle *Vehicle) RemoveBoard(board abstraction.Board) {
 	delete(vehicle.boards, board.Id())
 }
 
-func (vehicle *Vehicle) RemoveBoardId(id BoardId) {
+func (vehicle *Vehicle) RemoveBoardId(id abstraction.BoardId) {
 	delete(vehicle.boards, id)
 }
