+++ pkg/vehicle/vehicle.go
@@ -4,17 +4,6 @@ import (
 	"github.com/HyperloopUPV-H8/h9-backend/pkg/abstraction"
 )
 
-type BoardEvent string
-type BoardNotification interface {
-	Event() BoardEvent
-}
-
-type BoardId uint16
-type Board interface {
-	Id() BoardId
-	Notify(BoardNotification)
-}
-
 type LoggerName string
 type LoggerRecord interface {
 	Name() LoggerName
@@ -30,7 +19,7 @@ type Logger interface {
 
 type Vehicle struct {
 	broker    abstraction.Broker
-	boards    map[BoardId]Board
+	boards    map[abstraction.BoardId]abstraction.Board
 	transport abstraction.Transport
 	logger    Logger
 }
